update document storage adapter which includes always skipping certificate validation
