+++ medical-portal/src/API/RSBC.DMF.MedicalPortal.API/ServiceCollectionExtensions.cs
@@ -1,10 +1,8 @@
 ﻿using Grpc.Net.Client;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
+using Pssg.DocumentStorageAdapter;
 using Rsbc.Dmf.CaseManagement.Service;
 using Serilog;
 using System.Net;
-using System.Net.Http;
 
 namespace RSBC.DMF.MedicalPortal.API
 {
@@ -43,11 +41,11 @@ namespace RSBC.DMF.MedicalPortal.API
 
                     var initialClient = new CaseManager.CaseManagerClient(initialChannel);
                     // call the token service to get a token.
-                    var tokenRequest = new TokenRequest { Secret = clientSecret };
+                    var tokenRequest = new Rsbc.Dmf.CaseManagement.Service.TokenRequest { Secret = clientSecret };
 
                     var tokenReply = initialClient.GetToken(tokenRequest);
 
-                    if (tokenReply != null && tokenReply.ResultStatus == ResultStatus.Success)
+                    if (tokenReply != null && tokenReply.ResultStatus == Rsbc.Dmf.CaseManagement.Service.ResultStatus.Success)
                     {
                         // Add the bearer token to the client.
                         httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {tokenReply.Token}");
@@ -63,9 +61,46 @@ namespace RSBC.DMF.MedicalPortal.API
                         Log.Logger.Information("Error getting token for Case Management Service");
                     }
                 }
-
             }
             return services;
+        }
+
+        public static void AddDocumentStorageClient(this IServiceCollection services, IConfiguration configuration)
+        {
+            var documentStorageAdapterURI = configuration["DOCUMENT_STORAGE_ADAPTER_URI"];
+            if (!string.IsNullOrEmpty(documentStorageAdapterURI))
+            {
+                var httpClientHandler = new HttpClientHandler();
+
+                // Return `true` to allow certificates that are untrusted/invalid                    
+                httpClientHandler.ServerCertificateCustomValidationCallback =
+                    HttpClientHandler.DangerousAcceptAnyServerCertificateValidator;
+
+                var httpClient = new HttpClient(httpClientHandler);
+                // set default request version to HTTP 2.  Note that Dotnet Core does not currently respect this setting for all requests.
+                httpClient.DefaultRequestVersion = HttpVersion.Version20;
+
+                var initialChannel = GrpcChannel.ForAddress(documentStorageAdapterURI, new GrpcChannelOptions { HttpClient = httpClient, MaxReceiveMessageSize = null, MaxSendMessageSize = null });
+
+                var initialClient = new DocumentStorageAdapter.DocumentStorageAdapterClient(initialChannel);
+                // call the token service to get a token.
+                var tokenRequest = new Pssg.DocumentStorageAdapter.TokenRequest
+                {
+                    Secret = configuration["DOCUMENT_STORAGE_ADAPTER_JWT_SECRET"]
+                };
+
+                var tokenReply = initialClient.GetToken(tokenRequest);
+
+                if (tokenReply != null && tokenReply.ResultStatus == Pssg.DocumentStorageAdapter.ResultStatus.Success)
+                {
+                    // Add the bearer token to the client.
+                    httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {tokenReply.Token}");
+
+                    var channel = GrpcChannel.ForAddress(documentStorageAdapterURI, new GrpcChannelOptions { HttpClient = httpClient, MaxReceiveMessageSize = null, MaxSendMessageSize = null });
+
+                    services.AddTransient(_ => new DocumentStorageAdapter.DocumentStorageAdapterClient(channel));
+                }
+            }
         }
     }
 }
