+++ medical-portal/src/API/RSBC.DMF.MedicalPortal.API/Startup.cs
@@ -133,44 +133,7 @@ namespace RSBC.DMF.MedicalPortal.API
 
             // Add Document Storage Adapter
 
-            string documentStorageAdapterURI = configuration["DOCUMENT_STORAGE_ADAPTER_URI"];
-
-            if (!string.IsNullOrEmpty(documentStorageAdapterURI))
-            {
-                var httpClientHandler = new HttpClientHandler();
-                if (!environment.IsProduction()) // Ignore certificate errors in non-production modes.  
-                                                 // This allows you to use OpenShift self-signed certificates for testing.
-                {
-                    // Return `true` to allow certificates that are untrusted/invalid                    
-                    httpClientHandler.ServerCertificateCustomValidationCallback =
-                        HttpClientHandler.DangerousAcceptAnyServerCertificateValidator;
-                }
-
-                var httpClient = new HttpClient(httpClientHandler);
-                // set default request version to HTTP 2.  Note that Dotnet Core does not currently respect this setting for all requests.
-                httpClient.DefaultRequestVersion = HttpVersion.Version20;
-
-                var initialChannel = GrpcChannel.ForAddress(documentStorageAdapterURI, new GrpcChannelOptions { HttpClient = httpClient, MaxReceiveMessageSize = null, MaxSendMessageSize = null });
-
-                var initialClient = new DocumentStorageAdapter.DocumentStorageAdapterClient(initialChannel);
-                // call the token service to get a token.
-                var tokenRequest = new Pssg.DocumentStorageAdapter.TokenRequest
-                {
-                    Secret = configuration["DOCUMENT_STORAGE_ADAPTER_JWT_SECRET"]
-                };
-
-                var tokenReply = initialClient.GetToken(tokenRequest);
-
-                if (tokenReply != null && tokenReply.ResultStatus == Pssg.DocumentStorageAdapter.ResultStatus.Success)
-                {
-                    // Add the bearer token to the client.
-                    httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {tokenReply.Token}");
-
-                    var channel = GrpcChannel.ForAddress(documentStorageAdapterURI, new GrpcChannelOptions { HttpClient = httpClient, MaxReceiveMessageSize = null, MaxSendMessageSize = null });
-
-                    services.AddTransient(_ => new DocumentStorageAdapter.DocumentStorageAdapterClient(channel));
-                }
-            }
+            services.AddDocumentStorageClient(configuration);
 
             services.AddTransient<ICaseQueryService, CaseService>();
             services.AddTransient<IUserService, UserService>();
